fix: move get env variable to utils
