+++ packages/api/src/index.ts
@@ -1,4 +1,4 @@
-import { getEnvVariable } from '@etimo-achievements/common';
+import { getEnvVariable } from '@etimo-achievements/utils';
 import Server from './server';
 import { setupEnvironment } from './utils/setup-helper';
