+++ packages/common/src/env-helper.ts
@@ -1,6 +1,3 @@
-import { EnvVariable } from '@etimo-achievements/types';
-import { ConfigurationError } from '.';
-
 export enum Environment {
   Local = 'local',
   Development = 'development',
@@ -41,17 +38,4 @@ export function isStaging(): boolean {
 
 export function isProduction(): boolean {
   return getEnvironment() === Environment.Production;
-}
-
-export function getEnvVariable(name: EnvVariable, defaultValue?: string): string {
-  const value = process.env[name];
-  if (!value) {
-    if (defaultValue) {
-      console.debug(`Requested non-existant env variable ${name}, using default value: ${defaultValue}`);
-      return defaultValue;
-    }
-
-    throw new ConfigurationError(`Environment variable ${name} is not defined`);
-  }
-  return value;
 }
