+++ packages/utils/src/index.ts
@@ -1,3 +1,4 @@
 export * from './chat';
+export * from './env';
 export * from './feature';
 export * from './logger';
